done error handling

+++ 10_making-file-storage-app-in-express-server/routes/fileroutes.js
@@ -30,6 +30,12 @@ router.get("/:id", async (req, res, next) => {
 
         const file = fileData.files.find((file) => file.id === id);
 
+        if (!file) {
+            return res.status(404).json({
+                message: "File not found."
+            });
+        }
+
         if (req.query.action === "download") {
             res.set("Content-Disposition", `attachment; filename=${file.fileName}`);
         }
@@ -37,85 +43,94 @@ router.get("/:id", async (req, res, next) => {
         const fileName = path.join('/', file.id + file.ext);
         const filePath = path.join(path.resolve(import.meta.dirname, '..'), "storage", fileName);
 
-        res.sendFile(filePath);
+        return res.status(200).sendFile(filePath);
 
     } catch (error) {
         console.log(error);
+        next(error);
     }
 
 });
 
 // rename file
 router.patch("/:id", async (req, res, next) => {
+
+    const id = req.params.id;
+    console.log(id);
+    const newFileName = req.body.newFilename;
+
+    console.log(newFileName);
+
+    fileData.files.find((file) => file.id === id).fileName = newFileName;
+
     try {
-        const id = req.params.id;
-        console.log(id);
-        const newFileName = req.body.newFilename;
-
-        console.log(newFileName);
-
-        fileData.files.find((file) => file.id === id).fileName = newFileName;
-
         await writeFile("./utils/filesdata.json", JSON.stringify(fileData));
-
-        res.json({
+        return res.status(200).json({
             message: "File renamed successfully"
         });
 
     } catch (error) {
         console.log(error);
+        next(error);
     }
 });
 
 // delete file
 router.delete("/:id", async (req, res, next) => {
+
+    const id = req.params.id;
+    const file = fileData.files.find((file) => file.id === id);
+
+    if (!file) {
+        return res.status(404).json({
+            message: "File not found"
+        });
+    }
+
+    const filePath = path.join('/', file.id + file.ext);
+
+    fileData.files = fileData.files.filter((file) => file.id !== id);
+
+    const fileDir = dirData.dirs.find((dir) => dir.id === file.dirId);
+
+    fileDir.files = fileDir.files.filter((fileId) => fileId !== id);
+
+    dirData.dirs.find((dir) => dir.id === fileDir.id).files = fileDir.files;
+
     try {
-        const id = req.params.id;
-        const file = fileData.files.find((file) => file.id === id);
-        console.log(file);
-        const filePath = path.join('/', file.id + file.ext);
-
-        fileData.files = fileData.files.filter((file) => file.id !== id);
-
-        const fileDir = dirData.dirs.find((dir) => dir.id === file.dirId);
-
-        fileDir.files = fileDir.files.filter((fileId) => fileId !== id);
-
-        dirData.dirs.find((dir) => dir.id === fileDir.id).files = fileDir.files;
-
         await fs.unlink(path.join(path.resolve(import.meta.dirname, '..'), "storage", filePath));
-
         await writeFile("./utils/filesdata.json", JSON.stringify(fileData));
-
         await writeFile("./utils/foldersdata.json", JSON.stringify(dirData));
 
-        res.status(200).json({
+        return res.status(200).json({
             message: "File deleted successfully"
         });
 
     } catch (error) {
         console.log(error);
+        next(error);
     }
 });
 
 // upload normal way
 router.post("/:filename", async (req, res, next) => {
+
+    let fileName = req.params.filename || "untitled";
+    const ext = path.extname(fileName);
+    const id = crypto.randomUUID();
+    const fullName = path.join('/', id + ext);
+    const dirId = req.headers.dirid || dirData.dirs[0].id;
+
+    dirData.dirs.find((dir) => dir.id === dirId).files.push(id);
+
+    fileData.files.push({
+        id,
+        ext,
+        fileName,
+        dirId,
+    });
+
     try {
-        let fileName = req.params.filename;
-        const ext = path.extname(fileName);
-        const id = crypto.randomUUID();
-        const fullName = path.join('/', id + ext);
-        const dirId = req.headers.dirid || dirData.dirs[0].id;
-
-        dirData.dirs.find((dir) => dir.id === dirId).files.push(id);
-
-        fileData.files.push({
-            id,
-            ext,
-            fileName,
-            dirId,
-        });
-
         await writeFile("./utils/filesdata.json", JSON.stringify(fileData));
 
         await writeFile("./utils/foldersdata.json", JSON.stringify(dirData));
@@ -125,13 +140,13 @@ router.post("/:filename", async (req, res, next) => {
         req.pipe(writeStream);
 
         req.on('end', () => {
-            res.status(200).json({
+            return res.status(201).json({
                 message: "File uploaded successfully"
             });
         });
     } catch (error) {
         console.error('Error uploading file:', error);
-        res.status(500).send('An error occurred while uploading the file.');
+        return res.status(500).send('An error occurred while uploading the file.');
     }
 })
